Fixed return type of match to be an integer to be easier to use

+++ Matchismo/Matchismo/Model/Card.h
@@ -15,6 +15,6 @@
 @property (nonatomic, getter = isMatched) BOOL matched;
 @property (nonatomic, getter = isChosen) BOOL chosen;
 
-- (NSInteger)match:(NSArray*)otherCards;
+- (int)match:(NSArray*)otherCards;
 
 @end
